pyproject.toml
[project]
name = "ocr"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiofiles>=25.1.0",
    "aiohttp>=3.13.3",
    "asyncio>=4.0.0",
    "azure-ai-formrecognizer>=3.3.3",
    "python-dotenv>=1.2.2",
    "tenacity>=9.1.4",
]
